kernel: supercalls: allow userspace to pull list entries
this way userspace can pull up kernel's umount list and deduce by itself.
this is a bit of a pointerwalking mess but this allows us to

1. avoid a kmalloc kernel side
2. avoid potential crashes kernel side
3. maintain api backwards compatibility
4. userspace can deduce that the feature is there (get list size first)
5. userspace can get the list of entries

this can also help denylist handlers to deduce stuff and for advanced users
to do shit.

This should help on concerns brought up by tiann#2950 (comment)

[ rsuntk: Expose kernel_umount variable ]

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>
Signed-off-by: rsuntk <rissu.ntk@gmail.com>